Streamline config methods somewhat.
